new common file created
