+++ blocks/contentfragment/contentfragment.js
@@ -130,3 +130,5 @@ function renderHelper(data, template, callBack) {
   });
   return dom.innerHTML;
 }
+
+
