+++ component-definition.json
@@ -550,6 +550,21 @@
               }
             }
           }
+        },
+        {
+          "title": "Emi & Eligibilty Calculator",
+          "id": "emiandeligiblitycalc",
+          "plugins": {
+            "xwalk": {
+              "page": {
+                "resourceType": "core/franklin/components/block/v1/block",
+                "template": {
+                  "name": "emiandeligiblitycalc",
+                  "model": "emiandeligiblitycalc"
+                }
+              }
+            }
+          }
         }
       ]
     }
