+++ component-filters.json
@@ -74,7 +74,8 @@
       "mycontainer",
       "keyfeatures",
       "embed",
-      "homeloancalculator"
+      "homeloancalculator",
+      "emiandeligiblitycalc"
     ]
   },
   {
